# pkg/app/api/docs/swagger.yaml
basePath: /api/v1
definitions:
  ErrorResponse:
    properties:
      code:
        example: Internal
        type: string
      error:
        example: original server error message
        type: string
      msg:
        example: internal server error
        type: string
      status:
        example: error
        type: string
      trace_id:
        example: 1234567890abcdef1234567890abcdef
        type: string
    type: object
  SuccessResponse:
    properties:
      data:
        type: object
      status:
        example: success
        type: string
    type: object
...
  models.ProtoAuthToken:
    properties:
      created_date:
        type: string
      token:
        type: string
      ttl:
        maximum: 94608000
        minimum: 1
        type: integer
    required:
    - token
    - ttl
    type: object
  models.ProtoAuthTokenRequest:
    properties:
      ttl:
        default: 31536000
        maximum: 94608000
        minimum: 1
        type: integer
      type:
        default: browser
        enum:
        - aggregate
        - browser
        - external
        type: string
    required:
    - ttl
    - type
    type: object
...
info:
  contact:
    email: admin@vxcontrol.com
    name: Dmitry Nagibin
    url: https://vxcontrol.com
  description: Swagger API for VXControl SOLDR backend product.
  license:
    name: Proprietary License
    url: https://raw.githubusercontent.com/vxcontrol/soldr/master/LICENSE
  termsOfService: http://swagger.io/terms/
  title: SOLDR Swagger API
  version: "1.0"
paths:
...
  /export/modules/{module_name}/versions/{version}:
    post:
      parameters:
      - description: module name without spaces
        in: path
        name: module_name
        required: true
        type: string
      - default: latest
        description: module version string according semantic version format or 'latest'
          or 'all'
        in: path
        name: version
        required: true
        type: string
      produces:
      - application/octet-stream
      - application/json
      responses:
        "200":
          description: system module archive file
          schema:
            type: file
        "403":
          description: exporting system module content not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: system module or version not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on exporting system module
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Export of zip archive which contains selected system module and versions
      tags:
      - Modules
      - Export
...
  /import/modules/{module_name}/versions/{version}:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: module name without spaces
        in: path
        name: module_name
        required: true
        type: string
      - default: all
        description: module version string according semantic version format or 'all'
        in: path
        name: version
        required: true
        type: string
      - default: true
        description: override system module files and records in global DB
        in: query
        name: rewrite
        required: true
        type: boolean
      - description: system module archive file
        in: formData
        name: archive
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: system module archive uploaded successful
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: bad format input system module archive
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: importing system module content not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: system module or version in archive not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on importing system module
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Import from zip archive which contains selected system module and versions
      tags:
      - Modules
      - Import
...
  /token/vxproto:
    post:
      consumes:
      - application/json
      parameters:
      - description: Proto auth token request JSON data
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/models.ProtoAuthTokenRequest'
      produces:
      - application/json
      responses:
        "201":
          description: token created successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ProtoAuthToken'
              type: object
        "400":
          description: invalid requested token info
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: creating token not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on creating token
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Create new JWT token to use it into vxproto connections
      tags:
      - Proto
...
swagger: "2.0"